agents/company_search_agent/requirements.txt
uvicorn
sqlalchemy
psycopg2-binary
google-adk==1.21.0
a2a-sdk==0.2.16